pyproject.toml
[project]
name = "devot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan Teofilovic",email = "ivan.teofilovic@proton.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "ruff (>=0.11.5,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
